+++ src/main/java/org/opensearch/neuralsearch/processor/combination/GeometricMeanScoreCombinationTechnique.java
@@ -11,6 +11,7 @@ import java.util.Set;
 import lombok.ToString;
 import org.opensearch.neuralsearch.processor.explain.ExplainableTechnique;
 
+import static org.opensearch.neuralsearch.processor.combination.ScoreCombinationUtil.PARAM_NAME_WEIGHTS;
 import static org.opensearch.neuralsearch.processor.explain.ExplainUtils.describeCombinationTechnique;
 
 /**
@@ -20,7 +21,6 @@ import static org.opensearch.neuralsearch.processor.explain.ExplainUtils.describ
 public class GeometricMeanScoreCombinationTechnique implements ScoreCombinationTechnique, ExplainableTechnique {
     @ToString.Include
     public static final String TECHNIQUE_NAME = "geometric_mean";
-    public static final String PARAM_NAME_WEIGHTS = "weights";
     private static final Set<String> SUPPORTED_PARAMS = Set.of(PARAM_NAME_WEIGHTS);
     private static final Float ZERO_SCORE = 0.0f;
     private final List<Float> weights;