+++ src/main/java/org/opensearch/neuralsearch/processor/normalization/L2ScoreNormalizationTechnique.java
@@ -60,7 +60,7 @@ public class L2ScoreNormalizationTechnique implements ScoreNormalizationTechniqu
 
     @Override
     public String describe() {
-        return String.format(Locale.ROOT, "normalization technique [%s]", TECHNIQUE_NAME);
+        return String.format(Locale.ROOT, "normalization [%s]", TECHNIQUE_NAME);
     }
 
     @Override